backend/requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Database
alembic==1.12.1

# AI Services
google-generativeai==0.3.1
google-cloud-aiplatform==1.38.0
google-auth==2.23.4

# Image Processing
Pillow==10.1.0

# Frontend
streamlit==1.28.1
requests==2.31.0

# Data Science (for mock data and charts)
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
